#527 - Fixed renaming errors
