+++ src/containers/SearchPage/__tests__/_mockSearchResult.js
