+++ src/containers/SearchPage/__tests__/searchReducer-test.js
@@ -8,7 +8,7 @@
 
 import reducer, { initalState } from '../searchReducer';
 import * as actions from '../searchActions';
-import search from './_mockSearch';
+import searchResult from './_mockSearchResult';
 
 test('reducers/search initalState', () => {
   const nextState = reducer(undefined, { type: 'Noop' });
@@ -40,8 +40,8 @@ test('reducers/search searchError', () => {
 
 test('reducers/search handle set search result', () => {
   const nextState = reducer(initalState, {
-    type: actions.setSearch,
-    payload: search,
+    type: actions.setSearchResult,
+    payload: searchResult,
   });
 
   expect(nextState.totalCount).toBe(32);
@@ -52,8 +52,8 @@ test('reducers/search handle set search result', () => {
 });
 
 test('reducers/search handle clear search result', () => {
-  const nextState = reducer(search, {
-    type: actions.clearSearch,
+  const nextState = reducer(searchResult, {
+    type: actions.clearSearchResult,
   });
 
   expect(nextState).toEqual(initalState);
