--- src/containers/SearchPage/__tests__/searchResultReducer-test.js
@@ -1,60 +1,0 @@
-/**
- * Copyright (c) 2016-present, NDLA.
- *
- * This source code is licensed under the GPLv3 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-import reducer, { initalState } from '../searchReducer';
-import * as actions from '../searchActions';
-import search from './_mockSearch';
-
-test('reducers/search initalState', () => {
-  const nextState = reducer(undefined, { type: 'Noop' });
-
-  expect(nextState).toEqual({
-    results: [],
-    totalCount: 1,
-    pageSize: 10,
-    searching: false,
-  });
-});
-
-test('reducers/search search', () => {
-  const nextState = reducer(undefined, actions.search());
-
-  expect(nextState).toEqual({
-    results: [],
-    totalCount: 1,
-    pageSize: 10,
-    searching: true,
-  });
-});
-
-test('reducers/search searchError', () => {
-  const state = { ...initalState, searching: true };
-  const nextState = reducer(state, actions.searchError());
-  expect(nextState.searching).toBe(false);
-});
-
-test('reducers/search handle set search result', () => {
-  const nextState = reducer(initalState, {
-    type: actions.setSearch,
-    payload: search,
-  });
-
-  expect(nextState.totalCount).toBe(32);
-  expect(nextState.results.length).toBe(2);
-  expect(nextState.page).toBe(3);
-  expect(nextState.pageSize).toBe(2);
-  expect(nextState.searching).toBe(false);
-});
-
-test('reducers/search handle clear search result', () => {
-  const nextState = reducer(search, {
-    type: actions.clearSearch,
-  });
-
-  expect(nextState).toEqual(initalState);
-});
