+++ src/containers/SearchPage/searchSagas.js
@@ -16,7 +16,7 @@ export function* search(queryString) {
   try {
     const locale = yield select(getLocale);
     const searchResult = yield call(api.search, queryString, locale);
-    yield put(actions.setSearch(searchResult));
+    yield put(actions.setSearchResult(searchResult));
   } catch (error) {
     yield put(actions.searchError());
     // TODO: handle error
